+++ tests/unit/test_fields.py
@@ -5,20 +5,20 @@ import pytest
 import werkzeug.test
 import werkzeug.wrappers
 
-import flask_typed_routes.fields as flask_tpr_fields
+import flask_typed_routes.fields as ftr_fields
 
 
 def test_path_field(flask_app):
     with flask_app.test_request_context('/'):
         flask.request.view_args = {'id': 123}
-        assert flask_tpr_fields.Path(alias='id').value == 123
+        assert ftr_fields.Path(alias='id').value == 123
 
 
 @pytest.mark.parametrize('multi', [True, False])
 def test_query_field(flask_app, multi):
     expected = ['term'] if multi else 'term'
     with flask_app.test_request_context('/?search=term'):
-        result = flask_tpr_fields.Query(alias='search', multi=multi).value
+        result = ftr_fields.Query(alias='search', multi=multi).value
         assert result == expected
 
 
@@ -31,7 +31,7 @@ def test_cookie_field(flask_app, multi):
         env = builder.get_environ()
         request = werkzeug.wrappers.Request(env)
         with mock.patch('flask.request', request):
-            result = flask_tpr_fields.Cookie(alias='session_id', multi=multi).value
+            result = ftr_fields.Cookie(alias='session_id', multi=multi).value
 
         assert result == expected
 
@@ -43,17 +43,17 @@ def test_header_field(flask_app):
     ]
     # Fixed: The `headers` multi dont work
     with flask_app.test_request_context('/', headers=headers):
-        result = flask_tpr_fields.Header(alias='Authorization').value
+        result = ftr_fields.Header(alias='Authorization').value
         assert result == expected
 
 
 def test_body_field(flask_app):
     with flask_app.test_request_context('/', json={'key': 'value'}):
-        result = flask_tpr_fields.Body().value
+        result = ftr_fields.Body().value
     assert result == {'key': 'value'}
 
 
 def test_body_embed_field(flask_app):
     with flask_app.test_request_context('/', json={'key': {"subkey": "value"}}):
-        result = flask_tpr_fields.Body(alias="key", embed=True).value
+        result = ftr_fields.Body(alias="key", embed=True).value
     assert result == {"subkey": "value"}